replace pass template function with chezmoi data
Turns out, we don't need to use the pass template function to retrieve
trivial but somewhat sensitive data. We can store that inside the config
file of chezmoi and retrieve it from there. The pass template should
really only be used to retrieve access tokens and passwords stored on
the system in plain text, which I don't think exists on my system right
now.