use global git hooks to enforce commit-msg style
Until a few months ago, I was manually checking my commit message length
and making sure that the commit summary length didn't exceed *52* chars.
I'm not sure why I remembered *52* chars even though I had read this
blog post long ago.

https://tbaggery.com/2008/04/19/a-note-about-git-commit-messages.html

Anyways, this implements a global git hook for all the projects that I
have and enforces less than or equal to 50 characters for the commit
summary message and less than or equal to 72 characters for the commit
body message. It also checks that the 2nd line is empty or not.

If you're using Vim or Neovim, this happens automatically thanks to
tpope's code I guess but not if you use `git commit -m` on the CLI. This
commit-msg hook ensures that `git commit -m` conforms to this standard
as well.