disable trash-cli, add systemctl abbrev
In addition, the bind commands ensure garbage isn't printed when
ctrl+enter is pressed. Apparently, different terminals deal with this
differently. Alacritty and Kitty, by default, map ctrl+enter to enter.
However, zutty and foot don't and print garbage instead. In case of
foot, this was done to let the user change ctrl+enter to something else
if they want. I guess that's fair, although I almost always end up
pressing ctrl+enter in a hurry when working in a terminal, probably
because I've mapped my caps lock to behave as ctrl on hold and esc on
keypress and I use vim a lot.