eliminate layout shift when typing
I've disabled syntax highlighting in fish because of this issue

https://github.com/fish-shell/fish-shell/issues/8462

This was a source of annoyance for far too long and I tried to ignore it
but I couldn't take it anymore. I didn't realize this was an issue with
how terminal emulators render colored text. I tried using alacritty,
foot, kitty but all of them seem to suffer from this issue.