remove vscode specific config, fix onedark
I don't expect to use vscode anytime soon. Although neovim is mostly
inferior to vscode when it comes to IDE like features, it is better than
vscode when it comes to being an editor. In addition, after years of
using a window manager and a terminal, I can't tolerate high input
latency, which is where neovim is much better than vscode and always
will be.

I'll need to learn lua and fix issues that make me frustrated when using
neovim.